starters/pytorch-lightning/requirements.txt
pytorch_lightning
torchmetrics
torchvision
